Add files via upload
Explanation:
The 'createToken' function is public, so anyone can call it.
The 'transferToken' function is public, so anyone can call it. However, it has a require statement to ensure that the sender has a sufficient balance to make the transfer.
The 'getBalanceByToken' function is public, so anyone can call it.
The 'getTotalTokenCount' function is public, so anyone can call it.